# docs/swagger.yaml
definitions:
  api.BroadcastRequest:
    properties:
      message:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ServerInfo:
    properties:
      name:
        type: string
      version:
        type: string
    type: object
...
  api.ShutdownRequest:
    properties:
      message:
        type: string
      seconds:
        type: integer
    type: object
  api.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/server:
    get:
      consumes:
      - application/json
      description: Get Server Info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ServerInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get Server Info
      tags:
      - Server
  /api/server/broadcast:
    post:
      consumes:
      - application/json
      description: Publish Broadcast
      parameters:
      - description: Broadcast
        in: body
        name: broadcast
        required: true
        schema:
          $ref: '#/definitions/api.BroadcastRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Publish Broadcast
      tags:
      - Server
  /api/server/shutdown:
    post:
      consumes:
      - application/json
      description: Shutdown Server
      parameters:
      - description: Shutdown
        in: body
        name: shutdown
        required: true
        schema:
          $ref: '#/definitions/api.ShutdownRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Shutdown Server
      tags:
      - Server
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"